today()

+++ pyxcp/daq_stim/__init__.py
@@ -212,7 +212,7 @@ class DaqToCsv(DaqOnlinePolicy):
             hdr = ",".join(["timestamp0", "timestamp1"] + [h[0] for h in daq_list.headers])
             out_file.write(f"{hdr}\n")
 
-    def on_daq_list(self, daq_list: int, ts0: float, ts1: float, payload: list):
+    def on_daq_list(self, daq_list: int, ts0: int, ts1: int, payload: list):
         self.files[daq_list].write(f"{ts0},{ts1},{', '.join([str(x) for x in payload])}\n")
 
     def finalize(self):
